+++ routes/company.mjs
@@ -1,11 +1,11 @@
 import express from 'express'
 
-import {getAll, findById, save, actualize} from '../controllers/controller-company.js'
+import {getAll, getById, save, actualize} from '../controllers/controller-company.js'
 
 const route = express.Router()
 
 route.get('/',getAll)
-route.get('/:id',findById)
+route.get('/:id',getById)
 route.post('/',save)
 route.put('/:idn', actualize)
