+++ routes/products.mjs
@@ -2,7 +2,7 @@ import express from 'express'
 import {
     getAll,
     save,
-} from './../controllers/controller.product.js'
+} from '../controllers/controller-product.js'
 
 const routes = express.Router()
